+++ src/main/java/Arcadia/ClexaGod/arcadia/storage/PostgresConfig.java
@@ -1,5 +1,10 @@
 package Arcadia.ClexaGod.arcadia.storage;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
 public final class PostgresConfig {
 
     private final String host;
@@ -12,60 +17,6 @@ public final class PostgresConfig {
     private final int minIdle;
     private final int connectionTimeoutMs;
     private final int idleTimeoutMs;
-
-    public PostgresConfig(String host, int port, String database, String username, String password, boolean ssl,
-                          int maxPoolSize, int minIdle, int connectionTimeoutMs, int idleTimeoutMs) {
-        this.host = host;
-        this.port = port;
-        this.database = database;
-        this.username = username;
-        this.password = password;
-        this.ssl = ssl;
-        this.maxPoolSize = maxPoolSize;
-        this.minIdle = minIdle;
-        this.connectionTimeoutMs = connectionTimeoutMs;
-        this.idleTimeoutMs = idleTimeoutMs;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
-    public int getPort() {
-        return port;
-    }
-
-    public String getDatabase() {
-        return database;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public boolean isSsl() {
-        return ssl;
-    }
-
-    public int getMaxPoolSize() {
-        return maxPoolSize;
-    }
-
-    public int getMinIdle() {
-        return minIdle;
-    }
-
-    public int getConnectionTimeoutMs() {
-        return connectionTimeoutMs;
-    }
-
-    public int getIdleTimeoutMs() {
-        return idleTimeoutMs;
-    }
 
     public boolean isValid() {
         return host != null && !host.isBlank()